pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "formula-student-acceleration"
version = "1.0.0"
description = "Formula Student acceleration event simulation and optimization"
requires-python = ">=3.7"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
]

[tool.setuptools]
packages = ["config", "vehicle", "dynamics", "simulation", "rules", "analysis"]

